Refactor function to utils
